Moved version.py into standalone module
