+++ tests/kafkatest/services/performance/consumer_performance.py
@@ -14,11 +14,11 @@
 # limitations under the License.
 
 
+import os
+
 from kafkatest.services.performance import PerformanceService
 from kafkatest.services.security.security_config import SecurityConfig
-from kafkatest.version.version import TRUNK, V_0_9_0_0
-
-import os
+from kafkatest.version import TRUNK, V_0_9_0_0
 
 
 class ConsumerPerformanceService(PerformanceService):
