+++ tests/kafkatest/services/performance/producer_performance.py
@@ -13,16 +13,16 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+import os
+import subprocess
+
 from ducktape.utils.util import wait_until
 
+from kafkatest.directory_layout.kafka_path import  TOOLS_JAR_NAME, TOOLS_DEPENDANT_TEST_LIBS_JAR_NAME
 from kafkatest.services.monitor.jmx import JmxMixin
 from kafkatest.services.performance import PerformanceService
 from kafkatest.services.security.security_config import SecurityConfig
-from kafkatest.version.version import TRUNK, V_0_9_0_0
-from kafkatest.directory_layout.kafka_path import  TOOLS_JAR_NAME, TOOLS_DEPENDANT_TEST_LIBS_JAR_NAME
-
-import os
-import subprocess
+from kafkatest.version import TRUNK, V_0_9_0_0
 
 
 class ProducerPerformanceService(JmxMixin, PerformanceService):
