+++ tests/kafkatest/version.py
